+++ EasyBus/Migrations/20210323144147_AddedStopsModel.cs
@@ -16,7 +16,7 @@ namespace EasyBus.Migrations
                 name: "Stops",
                 columns: table => new
                 {
-                    Id = table.Column<long>(type: "bigint", nullable: false)
+                    Id = table.Column<long>(type: "int", nullable: false)
                         .Annotation("SqlServer:Identity", "1, 1"),
                     Name = table.Column<string>(type: "nvarchar(100)", maxLength: 100, nullable: false)
                 },
